requirements.txt
flask==3.1.*
requests==2.32.*
beautifulsoup4==4.13.*
newspaper3k==0.2.*
lxml_html_clean==0.4.*
scikit-learn==1.7.*
numpy==2.3.*
